requirements.txt
Django==4.2
djangorestframework==3.14.0
python-decouple==3.8
django_filter==23.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.5
psycopg2-binary==2.9.6
python-dateutil==2.8.2
gunicorn==21.0.1
